Move folderHasChildren up the hierarchy
